requirements.txt
nltk
spacy
wordcloud
matplotlib
rich
python-docx
pysub-parser
wikipedia
sphinxcontrib-programoutput
sphinx-autodoc-typehints
sphinx_rtd_theme